+++ src/compositor/mod.rs
@@ -2,5 +2,3 @@ pub mod dcomp;
 pub mod dx11;
 pub mod dx12;
 pub mod dx9;
-
-pub mod dx12_new;
