+++ src/hooks/mod.rs
@@ -14,10 +14,10 @@ use windows::Win32::UI::WindowsAndMessaging::{
     WS_EX_OVERLAPPEDWINDOW, WS_OVERLAPPEDWINDOW,
 };
 
-pub mod dx11;
+// pub mod dx11;
 pub mod dx12;
-pub mod dx9;
-pub mod opengl3;
+// pub mod dx9;
+// pub mod opengl3;
 
 /// A utility function to retrieve the top level [`HWND`] belonging to this
 /// process.
